08_maintain/requirements.txt
aiohttp
asyncio
google-cloud-pubsub
google-cloud-storage